test(client): reorganize tests structure
